Rename spt3g_2022. Add TT, EE, TE likelihoods
