+++ spt3g_2022/EE.yaml
@@ -1,0 +1,257 @@
+path: null
+
+# Bandpower file
+bandpower_filename: SPT3G_2018_TTTEEE_bandpowers.dat
+
+# Covariance file
+covariance_filename: SPT3G_2018_TTTEEE_covariance.dat
+
+# Fiducial Covariance file
+fiducial_covariance_filename: SPT3G_2018_TTTEEE_fiducial_covariance.dat
+
+# Beam covariance file
+beam_covariance_filename: SPT3G_2018_TTTEEE_beam_covariance.dat
+
+# Calibration (mapT, mapP) covariance
+cal_covariance_filename: SPT3G_2018_TTTEEE_cal_covariance.dat
+
+# effective frequency
+nu_eff_filename: SPT3G_2018_TTTEEE_effective_band_centres.dat
+
+# Windows directory
+window_folder: windows
+
+# Spectra to use (order does not matter)
+spectra_to_fit:
+  - 90_Ex90_E
+  - 90_Ex150_E
+  - 90_Ex220_E
+  - 150_Ex150_E
+  - 150_Ex220_E
+  - 220_Ex220_E
+
+foregrounds:
+  tSZ_template_filename: foreground_templates/tSZ_Dl_shaw10_153ghz_norm1.txt
+  kSZ_template_filename: foreground_templates/kSZ_Dl_CSF_incl_patchy_norm1.txt
+  galdust_T: 19.6
+  galdust_nu0: 150.0
+  CIB_T: 25.0
+  CIB_nu0: 150.0
+  tSZ_nu0: 143.0
+
+params:
+  kappa:
+    prior:
+      dist: norm
+      loc: 0.0
+      scale: 0.00045
+    ref: 0.0
+    proposal: 0.00045
+    latex: \kappa  
+
+  Tcal90:
+    value: 1.0
+    latex: T_{\rm{cal}^{90}}
+
+  Tcal150:
+    value: 0.9975
+    latex: T_{\rm{cal}^{150}}
+
+  Tcal220:
+    value: 0.9930
+    latex: T_{\rm{cal}^{220}}
+
+  Ecal90:
+    prior:
+      dist: norm
+      loc: 1.0
+      scale: 0.0087
+    ref: 1.0009
+    proposal: 0.01
+    latex: P_{\rm{cal}^{90}}
+
+  Ecal150:
+    prior:
+      dist: norm
+      loc: 1.0
+      scale: 0.0082
+    ref: 1.0020
+    proposal: 0.01
+    latex: P_{\rm{cal}^{150}}
+
+  Ecal220:
+    prior:
+      dist: norm
+      loc: 1.0
+      scale: 0.016
+    ref: 1.019
+    proposal: 0.01
+    latex: P_{\rm{cal}^{220}}
+
+  EE_Poisson_90x90:
+    prior:
+      dist: norm
+      loc: 0.041
+      scale: 0.012
+    ref: 0.041
+    proposal: 0.01
+    latex: D_{3000}^{\rm{PS,EE};90x90}
+
+  EE_Poisson_90x150:
+    prior:
+      dist: norm
+      loc: 0.0180
+      scale: 0.0054
+    ref: 0.0177
+    proposal: 0.01
+    latex: D_{3000}^{\rm{PS,EE};90x150}
+
+  EE_Poisson_90x220:
+    prior:
+      dist: norm
+      loc: 0.0157
+      scale: 0.0047
+    ref: 0.0157
+    proposal: 0.01
+    latex: D_{3000}^{\rm{PS,EE};90x220}
+
+  EE_Poisson_150x150:
+    prior:
+      dist: norm
+      loc: 0.0115
+      scale: 0.0034
+    ref: 0.115
+    proposal: 0.01
+    latex: D_{3000}^{\rm{PS,EE};150x150}
+
+  EE_Poisson_150x220:
+    prior:
+      dist: norm
+      loc: 0.0190
+      scale: 0.0057
+    ref: 0.0188
+    proposal: 0.01
+    latex: D_{3000}^{\rm{PS,EE};150x220}
+
+  EE_Poisson_220x220:
+    prior:
+      dist: norm
+      loc: 0.048
+      scale: 0.014
+    ref: 0.048
+    proposal: 0.01
+    latex: D_{3000}^{\rm{PS,EE};220x220}
+
+  EE_PolGalDust_Amp:
+    prior:
+      dist: norm
+      loc: 0.05
+      scale: 0.022
+    ref: 0.052
+    proposal: 0.01
+    latex: A_{80}^{\rm{EE}}
+
+  EE_PolGalDust_Alpha:
+    prior:
+      dist: norm
+      loc: -2.42
+      scale: 0.04
+    ref: -2.42
+    proposal: 0.01
+    latex: \alpha^{\rm{EE}}
+
+  EE_PolGalDust_Beta:
+    prior:
+      dist: norm
+      loc: 1.51
+      scale: 0.04
+    ref: 1.51
+    proposal: 0.01
+    latex: \beta^{\rm{EE}}
+
+
+  TE_PolGalDust_Amp:
+    value: 0.138
+    latex: A_{80}^{\rm{TE}}
+
+  TE_PolGalDust_Alpha:
+    value: -2.42
+    latex: \alpha^{\rm{TE}}
+
+  TE_PolGalDust_Beta:
+    value: 1.51
+    latex: \beta^{\rm{TE}}
+
+  TT_Poisson_90x90:
+    value: 62.61
+    latex: D_{3000}^{\rm{PS,TT};90x90}
+
+  TT_Poisson_90x150:
+    value: 27.9
+    latex: D_{3000}^{\rm{PS,TT};90x150}
+
+  TT_Poisson_90x220:
+    value: 24.3
+    latex: D_{3000}^{\rm{PS,TT};90x220}
+
+  TT_Poisson_150x150:
+    value: 16.7
+    latex: D_{3000}^{\rm{PS,TT};150x150}
+
+  TT_Poisson_150x220:
+    value: 28.6
+    latex: D_{3000}^{\rm{PS,TT};150x220}
+
+  TT_Poisson_220x220:
+    value: 78.5
+    latex: D_{3000}^{\rm{PS,TT};220x220}
+
+  TT_GalCirrus_Amp:
+    value: 1.93
+    latex: A^{\rm{cirrus}}_{80}
+
+  TT_GalCirrus_Alpha:
+    value: -2.53
+    latex: \alpha^{\rm{cirrus}}
+
+  TT_GalCirrus_Beta:
+    value: 1.48
+    latex: \beta^{\rm{cirrus}}
+
+  TT_CIBClustering_Amp:
+    value: 5.2
+    latex: A^{\rm{CIB}-cl}_{80}
+
+  TT_CIBClustering_Alpha:
+    value: 0.8
+    latex: \alpha^{\rm{CIB}-cl}
+
+  TT_CIBClustering_Beta:
+    value: 0.1
+    latex: \beta^{\rm{CIB}-cl}
+
+  TT_CIBClustering_decorr_90:
+    value: 1.
+    latex: \xi^{\rm{CIB}-decorr,90}
+
+  TT_CIBClustering_decorr_150:
+    value: 1.
+    latex: \xi^{\rm{CIB}-decorr,150}
+
+  TT_CIBClustering_decorr_220:
+    value: 1.
+    latex: \xi^{\rm{CIB}-decorr,220}
+
+  TT_tSZ_Amp:
+    value: 4.7
+    latex: A_{\rm{tSZ}}
+
+  TT_tSZ_CIB_corr:
+    value: 0.09
+    latex: \xi
+
+  TT_kSZ_Amp:
+    value: 3.7
+    latex: A_{\rm{kSZ}}
+
+
