+++ spt3g_2022/__init__.py
@@ -1,1 +1,1 @@
-from .spt3g_TTTEEE import TTTEEE
+from .spt3g_2022 import TTTEEE, TT, TE, EE
