+++ spt3g_2022/spt3g_2022_foregrounds.py
